Vmware: Move session test to down module
VMwareAPISession has been moved to its own module,
and this change should reflect that in the test case.

(cherry picked from commit 98541682995e5dd8838e5714964fb4be25593aa3)
